update number type to BigDecimal

+++ src/main/java/com/huobi/api/client/domain/Merged.java
@@ -3,6 +3,8 @@ package com.huobi.api.client.domain;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 /**
  * created by jacky. 2018/7/21 2:12 PM
  */
@@ -13,13 +15,13 @@ public class Merged {
      * K线id
      */
     private String id;
-    private double open;
-    private double close;
-    private double low;
-    private double high;
-    private double amount;
-    private double vol;
-    private int count;
+    private BigDecimal open;
+    private BigDecimal close;
+    private BigDecimal low;
+    private BigDecimal high;
+    private BigDecimal amount;
+    private BigDecimal vol;
+    private BigDecimal count;
     private String symbol;
     /**
      *[买1价,买1量],
