+++ src/main/java/com/huobi/api/client/domain/TickerPrice.java
@@ -3,6 +3,8 @@ package com.huobi.api.client.domain;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 /**
  * created by jacky. 2018/7/21 2:17 PM
  */
@@ -10,6 +12,6 @@ import lombok.Setter;
 @Setter
 public class TickerPrice {
    private String symbol;
-   private double price;
-   private double vol;
+   private BigDecimal price;
+   private BigDecimal vol;
 }
